refactor: merge type folders
